Moved apparatus data along with  ugnt text card, Fixed responsiveness
