+++ jsconfig.json
@@ -4,6 +4,7 @@
     "paths": {
       "@components/*": ["src/components/*"],
       "@context/*": ["src/context/*"],
+      "@utils/*": ["src/utils/*"],
       "@styles/*": ["styles/*"]
     }
   }
