+++ src/utils/filterApparatusReadings.js
@@ -1,0 +1,31 @@
+export default function filterApparatusReadings({
+  variantObjects,
+  requireIsTranslatable,
+  requireAncient,
+}) {
+  let filteredVariantObjects = []
+
+  variantObjects.forEach(currentVariantObject => {
+    const filteredReadings = currentVariantObject.readings.filter(
+      vo => vo.isTranslatable || requireIsTranslatable == false
+    )
+
+    currentVariantObject.readings = []
+    filteredReadings.forEach(rd => {
+      // Filter sources:
+      rd.sources = rd.sources.filter(
+        src => src.textClass !== 'mod' || requireAncient == false
+      )
+
+      if (rd.sources.length > 0) {
+        currentVariantObject.readings.push(rd)
+      }
+    })
+
+    if (currentVariantObject.readings.length > 0) {
+      filteredVariantObjects.push(currentVariantObject)
+    }
+  })
+
+  return filteredVariantObjects
+}
